Move cell model definition to JSON file
This sets us up to have multiple model definitions.

Since the JSON file doesn't have access to the view class, the properties
such as `modelProperties` and `elId` have to be added after the fact.

Removed the in-JSON `clickHandlers` definitions and replace them with
equivalent `model.on('view.click', ...)` definitions.

Had to add a module declaration for `"*.json"` to allow importing JSON
files as a module.

[#155415014]