Keep substance at raised level for few seconds
This adds the "box-shaped" graph, whereby a user's manipulations persist
for some time before reverting back to the original value.

The `step` function now passes in the "real" time from the beginning of
the simulation.

Substance now has a new `manuallyIncrement` function, which adds an end-
time to when the substance manipulation should persist. This end-time will
keep getting pushed back if the user clicks multiple times.

Before the end time is reached, there will be no birth or death rate for
that substance, so it will flat-line (unless the user manually changes it
again).

[#155597017]