Update cell model properties every step
Previously we were updating the Organism model every time tick (to set
the substance levels), but the cell model's `modelProperties` were only
triggering a change on certain prop changes. This may have been because
of the nested nature of `props.organism.modelProperties`.

In any case, every model tick we now request the modelProperties and
update the Organelle model.