Add BioLogica Org to Organism Model
Add BioLogica as a library. Require Organisms to be created with an
allele string. Add a view to the model, `getBaseColor`, which returns
`Light` or `Dark` given the underlying organism genetics. We can then
use this anywhere that we were originally hard-coding a look up to the
`Field Mouse` and `Beach Mouse` ids.

Create our default `Field Mouse` and `Beach Mouse` organisms using
genetics.