customOrgs url param; move availableOrgs to rootStore
This adds a new url parameter, `customOrgs`, which can be used to define
any number of additional organisms by allele string, e.g.

    ?customOrgs=a:B,b:B;a:b,b:b

This will then create the organisms and add them to both the list of
organisms and the list of availableOrgs.

Moved the availableOrgs list to the rootStore to better handle this,
so that the same store is in charge of what orgs there are and what
orgs are available. Otherwise it was complicated to arrange the order
of creation between the different stores.