Add geniblocks views; add "hidden" Genome view
This adds the GeniBlocks library, and a WIP version of a genome view,
which can be found by adding `Genome` to the `availableViews` url
parameter. By default, it is not one of the available views.

    ?availableViews=Organism,Cell,Protein,Genome

with the addition of another url parameter, `canEditGenome`, the alleles
of any organism are editable:

    ?availableViews=Organism,Cell,Protein,Genome&canEditGenome=true

Changing the genome sets the `alleles` of an organism, which automagically
changes the underlying `biologicaOrganism`, the `baseColor`, and therefore
the melanin substances `step` function, which will cause the organism to
lighten several seconds after the alleles are changed.